more requested changes done refactor folder name removed unused dictionary ( to be used in the future)
