+++ src/app/[lang]/games/memo-blocks/page.tsx
@@ -1,12 +1,7 @@
-
 import { LangParam } from '@/types';
-import { getDictionaryServer } from '@/utils/getDictionaryServer';
-import MemoBlocksGame from '@/components/shared/memoBlocks';
+import MemoBlocksGame from '@/components/shared/memoBlocksGame';
 
 const Page = async ({ params: { lang } }: LangParam) => {
-    const dict = await getDictionaryServer(lang);
-    const { page } = dict.app.contact;
-
     return (
         <div className='h-full flex justify-center items-center'>
             <MemoBlocksGame />
