+++ src/components/shared/memoBlocksGame/SortableCard.tsx
